Promoted fps and delays to doubles

+++ src/main/java/animtool/animation/Frame.java
@@ -26,10 +26,7 @@ package animtool.animation;
 
 
 import animtool.gui.Main;
-import javafx.beans.property.IntegerProperty;
-import javafx.beans.property.ObjectProperty;
-import javafx.beans.property.SimpleIntegerProperty;
-import javafx.beans.property.SimpleObjectProperty;
+import javafx.beans.property.*;
 import javafx.scene.image.Image;
 
 import java.io.File;
@@ -42,14 +39,14 @@ public class Frame implements Comparable<Frame> {
     public static final int THUMBNAIL_SIZE = 100;
 
     private final File file;
-    private final IntegerProperty defaultDelay;
+    private final DoubleProperty defaultDelay;
 
     private final ObjectProperty<Image> image = new SimpleObjectProperty<>();
     private final ObjectProperty<Image> thumbnail = new SimpleObjectProperty<>();
-    private final IntegerProperty delay = new SimpleIntegerProperty(-1);
+    private final DoubleProperty delay = new SimpleDoubleProperty(-1);
 
 
-    public Frame(File file, IntegerProperty defaultDelay) {
+    public Frame(File file, DoubleProperty defaultDelay) {
         this.file = file;
         this.defaultDelay = defaultDelay;
     }
@@ -98,29 +95,29 @@ public class Frame implements Comparable<Frame> {
         return thumb;
     }
 
-    public synchronized int getDefaultDelay() {
+    public synchronized double getDefaultDelay() {
         return defaultDelay.get();
     }
 
-    public IntegerProperty defaultDelayProperty() {
+    public DoubleProperty defaultDelayProperty() {
         return defaultDelay;
     }
 
-    public synchronized int getDelay() {
+    public synchronized double getDelay() {
         return delay.get();
     }
 
-    public synchronized void setDelay(int delay) {
+    public synchronized void setDelay(double delay) {
         this.delay.set(delay);
     }
 
-    public synchronized int getComputedDelay() {
-        int d = getDelay();
+    public synchronized double getComputedDelay() {
+        double d = getDelay();
         if (d <= 0) d = getDefaultDelay();
         return d;
     }
 
-    public IntegerProperty delayProperty() {
+    public DoubleProperty delayProperty() {
         return delay;
     }
